requirements.txt
backports.functools-lru-cache==1.4
bottle==0.12.13
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.11.6
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-crontab==0.7.1
django-filter==1.0.4
django-filters==0.2.1
django-formtools==2.1
django-guardian==1.4.9
django-import-export==0.5.1
django-redis==4.8.0
django-reversion==2.0.10
djangorestframework==3.7.0
djangorestframework-jwt==1.11.0
DjangoUeditor==1.8.143
drf-extensions==0.3.1
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.10.3
idna==2.6
itchat==1.2.32
itypes==1.1.0
jdcal==1.3
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.1
mysqlclient==1.3.12
numpy==1.13.3
oauthlib==2.0.4
odfpy==1.3.5
olefile==0.44
opencv-python==3.4.0.12
openpyxl==2.4.8
optionaldict==0.1.1
Pillow==4.3.0
pycryptodome==3.4.7
PyJWT==1.5.3
PyMySQL==0.8.0
pyparsing==2.2.0
pypng==0.0.18
PyQRCode==1.2.1
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.3
PyYAML==3.12
raven==6.2.1
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
uWSGI==2.0.15
wechatpy==1.6.0
WeRoBot==1.2.0
wxpy==0.3.9.8
xlrd==1.1.0
XlsxWriter==1.0.0
xlwt==1.3.0
xmltodict==0.11.0
